+++ HISTORY.md
@@ -1,3 +1,12 @@
+Version 1.0
+===========
+June 14, 2017
+
++ Refactored WikipediaBz2DumpInputStream into WikipediaArticlesDump, which provides iterator and stream support
++ Refactored builder
++ Changed visibility scope of methods and inner classes of WikiClean, test cases refactored accordingly
++ Added documentation
+
 Version 0.5
 ===========
 March 25, 2017